+++ crates/onda_mir/src/format.rs
@@ -619,12 +619,18 @@ fn format_rvalue(value: &Rvalue) -> String {
         Rvalue::BufferSampleRate(buffer) => {
             format!("buffer_sample_rate {}", format_buffer_ref(*buffer))
         }
+        Rvalue::BufferIsBound(buffer) => {
+            format!("buffer_is_bound {}", format_buffer_ref(*buffer))
+        }
         Rvalue::BufferParamLen(id) => format!("buffer_len {}", format_buffer_param_ref(*id)),
         Rvalue::BufferParamChannels(id) => {
             format!("buffer_channels {}", format_buffer_param_ref(*id))
         }
         Rvalue::BufferParamSampleRate(id) => {
             format!("buffer_sample_rate {}", format_buffer_param_ref(*id))
+        }
+        Rvalue::BufferParamIsBound(id) => {
+            format!("buffer_is_bound {}", format_buffer_param_ref(*id))
         }
         Rvalue::ConstDataLoad {
             data,